+++ client/src/components/index.js
@@ -9,3 +9,4 @@ export { default as AddToCartButton } from './AddToCartBtn'
 export { default as AddToWishes } from './AddToWishBtn'
 export { default as SessionButton } from './SessionButton';
 export { default as CartCard } from './CartCard';
+export { default as Filter } from './Filters/filtro';
