+++ client/src/sections/index.js
@@ -1,4 +1,4 @@
-export { default as Home } from './Home';
+export { default as Filter } from './Filter';
 export { default as About } from './About';
 export { default as Games } from './Games';
 export { default as LandingFormSection } from './LandingFormSection';
